% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_usdos}
\alias{run_usdos}
\title{Runs USDOS model for a given config file.}
\usage{
run_usdos(cfile)
}
\arguments{
\item{cfile}{the name of the config file to use}
}
\value{
writes output to file
}
\description{
Runs USDOS model for a given config file.
}

