% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_usdos}
\alias{run_usdos}
\title{Runs USDOS model for a given config file.}
\usage{
run_usdos(cfile, gen_shipment_network = FALSE)
}
\arguments{
\item{cfile}{The name of the config file to use}

\item{gen_shipment_network}{Logical indicating whether you want to just generate shipments, defaults to FALSE}
}
\value{
writes output to file specified in cfile
}
\description{
Runs USDOS model for a given config file.
}
\examples{
run_usdos("config.txt")
}

